# api/api-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/athletes": {
      "get": {
        "operationId": "AthletesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List all athletes",
        "tags": [
          "athletes"
        ]
      },
      "post": {
        "operationId": "AthletesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAthleteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Athlete created"
          }
        },
        "summary": "Create a new athlete",
        "tags": [
          "athletes"
        ]
      }
    },
    "/athletes/{id}": {
      "get": {
        "operationId": "AthletesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get athlete by ID",
        "tags": [
          "athletes"
        ]
      },
      "put": {
        "operationId": "AthletesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAthleteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Athlete updated"
          }
        },
        "summary": "Update an existing athlete",
        "tags": [
          "athletes"
        ]
      },
      "delete": {
        "operationId": "AthletesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Athlete deleted"
          }
        },
        "summary": "Delete an athlete",
        "tags": [
          "athletes"
        ]
      }
    }
  },
  "info": {
    "title": "Private API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateAthleteDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "First name of the athlete",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "Last name of the athlete",
            "example": "Doe"
          },
          "countryCode": {
            "type": "string",
            "description": "ISO country code (e.g., FR, US)",
            "example": "FR"
          }
        },
        "required": [
          "firstName",
          "lastName"
        ]
      },
      "UpdateAthleteDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "First name of the athlete",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "Last name of the athlete",
            "example": "Doe"
          },
          "countryCode": {
            "type": "object",
            "description": "ISO country code (e.g., FR, US)",
            "example": "FR"
          }
        }
      }
    }
  }
}